fixed scroll animations on separate pages and changed tech stack layout for mobile device
