+++ src/pages/Home/Home.css
@@ -332,7 +332,7 @@
 
     .TechCard__Container {
         gap: 12px;
-        grid-template-columns: 1fr 1fr 1fr 1fr;
+        grid-template-columns: 1fr 1fr;
     }
 
     .ContactForm__Inputs {
